+++ my-app/database/migrations/2026_08_05_042830_create_applications_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedInteger('num')->unique();
             $table->string('company');
             $table->string('role');
-            $table->unsignedTinyInteger('score')->nullable();
+            $table->decimal('score', 3, 1)->nullable();
             $table->string('status')->nullable();
             $table->date('date')->nullable();
             $table->string('report_link')->nullable();
